+++ src/by/it/basumatarau/jd01_05/Main.java
@@ -1,0 +1,7 @@
+package by.it.basumatarau.jd01_05;
+
+public class Main {
+    public static void main(String[] args) {
+
+    }
+}
