+++ src/by/it/basumatarau/jd01_05/TaskA.java
@@ -1,0 +1,8 @@
+package by.it.basumatarau.jd01_05;
+
+public class TaskA {
+    public static void main(String[] args) {
+
+    }
+
+}
